pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "BookStore API"
authors = ["juan-mendes-DEV <jp436807@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"
asgiref = "^3.8.1"
colorama = { version = "^0.4.6", markers = "sys_platform == 'win32'" }
faker = "^33.1.0"
iniconfig = "^2.0.0"
packaging = "^24.2"
pluggy = "^1.5.0"
pytest = "^8.3.4"
python-dateutil = "^2.9.0.post0"
six = "^1.17.0"
sqlparse = "^0.5.3"
typing-extensions = "^4.12.2"
tzdata = { version = "^2024.2", markers = "sys_platform == 'win32'" }
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.8.2"
gunicorn = "^23.0.0"
django-environ = "^0.11.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
